Move Jira logic into separate package
