+++ issues/issues.go
@@ -21,13 +21,13 @@ import (
 	"strings"
 	"time"
 
-	"github.com/andygrunwald/go-jira"
+	gojira "github.com/andygrunwald/go-jira"
 	gh "github.com/google/go-github/v48/github"
 
 	"github.com/uwu-tools/gh-jira-issue-sync/cfg"
 	"github.com/uwu-tools/gh-jira-issue-sync/comments"
 	"github.com/uwu-tools/gh-jira-issue-sync/github"
-	"github.com/uwu-tools/gh-jira-issue-sync/lib/clients"
+	"github.com/uwu-tools/gh-jira-issue-sync/jira"
 )
 
 // dateFormat is the format used for the Last IS Update field.
@@ -37,7 +37,7 @@ const dateFormat = "2006-01-02T15:04:05.0-0700"
 // gets the list of JIRA issues which have GitHub ID custom fields in that list,
 // then matches each one. If a JIRA issue already exists for a given GitHub issue,
 // it calls UpdateIssue; if no JIRA issue already exists, it calls CreateIssue.
-func Compare(config cfg.Config, ghClient github.Client, jiraClient clients.JIRAClient) error {
+func Compare(config cfg.Config, ghClient github.Client, jiraClient jira.Client) error {
 	log := config.GetLogger()
 
 	log.Debug("Collecting issues")
@@ -92,7 +92,7 @@ func Compare(config cfg.Config, ghClient github.Client, jiraClient clients.JIRAC
 
 // DidIssueChange tests each of the relevant fields on the provided JIRA and GitHub issue
 // and returns whether or not they differ.
-func DidIssueChange(config cfg.Config, ghIssue gh.Issue, jIssue jira.Issue) bool {
+func DidIssueChange(config cfg.Config, ghIssue gh.Issue, jIssue gojira.Issue) bool {
 	log := config.GetLogger()
 
 	log.Debugf("Comparing GitHub issue #%d and JIRA issue %s", ghIssue.GetNumber(), jIssue.Key)
@@ -133,15 +133,15 @@ func DidIssueChange(config cfg.Config, ghIssue gh.Issue, jIssue jira.Issue) bool
 // UpdateIssue compares each field of a GitHub issue to a JIRA issue; if any of them
 // differ, the differing fields of the JIRA issue are updated to match the GitHub
 // issue.
-func UpdateIssue(config cfg.Config, ghIssue gh.Issue, jIssue jira.Issue, ghClient github.Client, jClient clients.JIRAClient) error {
+func UpdateIssue(config cfg.Config, ghIssue gh.Issue, jIssue gojira.Issue, ghClient github.Client, jClient jira.Client) error {
 	log := config.GetLogger()
 
 	log.Debugf("Updating JIRA %s with GitHub #%d", jIssue.Key, *ghIssue.Number)
 
-	var issue jira.Issue
+	var issue gojira.Issue
 
 	if DidIssueChange(config, ghIssue, jIssue) {
-		fields := jira.IssueFields{}
+		fields := gojira.IssueFields{}
 		fields.Unknowns = map[string]interface{}{}
 
 		fields.Summary = ghIssue.GetTitle()
@@ -159,7 +159,7 @@ func UpdateIssue(config cfg.Config, ghIssue gh.Issue, jIssue jira.Issue, ghClien
 
 		fields.Type = jIssue.Fields.Type
 
-		issue = jira.Issue{
+		issue = gojira.Issue{
 			Fields: &fields,
 			Key:    jIssue.Key,
 			ID:     jIssue.ID,
@@ -190,13 +190,13 @@ func UpdateIssue(config cfg.Config, ghIssue gh.Issue, jIssue jira.Issue, ghClien
 
 // CreateIssue generates a JIRA issue from the various fields on the given GitHub issue, then
 // sends it to the JIRA API.
-func CreateIssue(config cfg.Config, issue gh.Issue, ghClient github.Client, jClient clients.JIRAClient) error {
+func CreateIssue(config cfg.Config, issue gh.Issue, ghClient github.Client, jClient jira.Client) error {
 	log := config.GetLogger()
 
 	log.Debugf("Creating JIRA issue based on GitHub issue #%d", *issue.Number)
 
-	fields := jira.IssueFields{
-		Type: jira.IssueType{
+	fields := gojira.IssueFields{
+		Type: gojira.IssueType{
 			Name: "Task", // TODO: Determine issue type
 		},
 		Project:     config.GetProject(),
@@ -218,7 +218,7 @@ func CreateIssue(config cfg.Config, issue gh.Issue, ghClient github.Client, jCli
 
 	fields.Unknowns[config.GetFieldKey(cfg.LastISUpdate)] = time.Now().Format(dateFormat)
 
-	jIssue := jira.Issue{
+	jIssue := gojira.Issue{
 		Fields: &fields,
 	}
